Fixed save files in RaBF [v0.1]
When you load the game it will verify if the save files exist, and if it don't it would show this text on screen "You didn't started any games yet." but when you created save files, for some reason the game gave you that message, it was a weird bug but i was able to fix it.  [ it was a Hotfix, great! ]